Move TrackBuilder to Track folder
