+++ Tests/MusicLib/trackTest.cpp
@@ -2,7 +2,7 @@
 
 #include <MusicLib/Types/Track/TrackEvents/noteEvents.hpp>
 #include <MusicLib/Types/Track/track.hpp>
-#include <MusicLib/Utilities/trackBuilder.hpp>
+#include <MusicLib/Types/Track/trackBuilder.hpp>
 
 #include <Tests/TestUtils/seqTestUtils.hpp>
 #include <Tests/TestUtils/testTrackEvents.hpp>
